From http://www.jwz.org/xscreensaver/xscreensaver-5.43.tar.gz
-rw-rw-r-- 1 zblaxell zblaxell 26968319 Jul  2 16:27 xscreensaver-5.43.tar.gz
26f5f2535869ec90926e3e6f6df342cf96548c8f  xscreensaver-5.43.tar.gz

+++ hacks/glx/glhanoi.c
@@ -1296,7 +1296,7 @@ static void initData(glhcfg *glhanoi)
 		glhanoi->pole[i].size = glhanoi->numberOfDisks;
 	}
 	checkAllocAndExit(
-			!!(glhanoi->diskPos = calloc(glhanoi->numberOfDisks, sizeof(double))),
+			!!(glhanoi->diskPos = calloc(glhanoi->numberOfDisks, sizeof(float))),
 			"diskPos");
 
 	if (glhanoi->trailQSize) {
@@ -1430,7 +1430,7 @@ static GLubyte *makeTexture(glhcfg *glhanoi, int x_size, int y_size, int z_size,
 	double xi, yi, zi;
 
 	if((textureData =
-		calloc(x_size * y_size * z_size, sizeof(GLuint))) == NULL) {
+		calloc(x_size * y_size * z_size, sizeof(GLubyte))) == NULL) {
 		return NULL;
 	}
